SDIT-3127 Implement fix to delete method return type (#352)

+++ src/test/java/uk/gov/justice/digital/hmpps/cmd/api/service/NotificationServiceTest.kt
@@ -368,7 +368,7 @@ internal class NotificationServiceTest {
 
     @Test
     fun `Tidy notifications`() {
-      every { shiftNotificationRepository.deleteAllByShiftModifiedBefore(any()) } returns 4 as Integer
+      every { shiftNotificationRepository.deleteAllByShiftModifiedBefore(any()) } returns 4
 
       service.tidyNotification()
